requirements.txt
 Django==5.1.3
 djangorestframework==3.15.2